refactor : 위도, 경도 request 메소드 이동 (#106)
