+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/bakery/service/BakeryFavoritePageService.java
@@ -1,33 +1,31 @@
 package com.breaditnow.customer.domain.bakery.service;
 
+import com.breaditnow.customer.domain.bakery.controller.res.BakeryFavoritePageResponse;
+import com.breaditnow.customer.domain.common.req.GeoPointRequest;
+import com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.CustomerBakeryFavoriteRepository;
+import com.breaditnow.domain.global.dto.BakeryFavoriteDistanceDto;
+import com.breaditnow.domain.global.dto.GeoPoint;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.breaditnow.customer.domain.bakery.controller.req.GeoPointRequest;
-import com.breaditnow.customer.domain.bakery.controller.res.BakeryFavoritePageResponse;
-import com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.CustomerBakeryFavoriteRepository;
-import com.breaditnow.domain.global.dto.BakeryFavoriteDistanceDto;
-import com.breaditnow.domain.global.dto.GeoPoint;
-
-import lombok.RequiredArgsConstructor;
-
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class BakeryFavoritePageService {
-	private final CustomerBakeryFavoriteRepository customerBakeryFavoriteRepository;
+    private final CustomerBakeryFavoriteRepository customerBakeryFavoriteRepository;
 
-	public BakeryFavoritePageResponse getFavorites(Long customerId, int page, int size, String sort,
-		GeoPointRequest geoPointRequest) {
-		GeoPoint currentGeoPoint = geoPointRequest.toEntity();
-		Pageable pageable = PageRequest.of(page, size);
+    public BakeryFavoritePageResponse getFavorites(Long customerId, int page, int size, String sort,
+                                                   GeoPointRequest geoPointRequest) {
+        GeoPoint currentGeoPoint = geoPointRequest.toEntity();
+        Pageable pageable = PageRequest.of(page, size);
 
-		Page<BakeryFavoriteDistanceDto> pageBakeryFavorites = customerBakeryFavoriteRepository.findBakeryFavorites(
-			customerId, pageable, sort, currentGeoPoint);
+        Page<BakeryFavoriteDistanceDto> pageBakeryFavorites = customerBakeryFavoriteRepository.findBakeryFavorites(
+                customerId, pageable, sort, currentGeoPoint);
 
-		return BakeryFavoritePageResponse.of(pageBakeryFavorites);
-	}
+        return BakeryFavoritePageResponse.of(pageBakeryFavorites);
+    }
 }
