--- breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/region/controller/req/LocationRequest.java
@@ -1,6 +1,0 @@
-package com.breaditnow.customer.domain.region.controller.req;
-
-public record LocationRequest(double latitude, double longitude) {
-
-}
-
